Add different elements to add left and right side
Added a lot of text and rewrote a lot of things in
the index.html and style.css. Goal was to add an
edge/space on left and right side of all modules
on the landing page and still having all the
modules lined up horizontally.

It is better than before but not optimal.

I will keep working on that matter.