Add content div and headline under the hero part
Added the div for the main content for background
color change to white as well as an div inside with
the headline under the hero.

Info: The footer is still not being pushed down,
will deal with that after the content is added.
Probably it doesn't even have to be pushed down if
all content is added.